taskplanner v2 добавил H2 для тестов

+++ src/main/java/ru/otus/java/pro/taskplanner/DataLoader.java
@@ -7,6 +7,8 @@ import ru.otus.java.pro.taskplanner.model.User;
 import ru.otus.java.pro.taskplanner.repository.TaskRepository;
 import ru.otus.java.pro.taskplanner.repository.UserRepository;
 
+import java.util.Arrays;
+
 /**
  * Класс для наполнения синтетическими данными при старте приложения(просто для демонстрациии работы)
  */
@@ -23,37 +25,39 @@ public class DataLoader implements CommandLineRunner {
 
     @Override
     public void run(String... args) {
-        userRepository.deleteAll();
-        taskRepository.deleteAll();
+        if (!Arrays.asList(args).contains("test")) {
+            userRepository.deleteAll();
+            taskRepository.deleteAll();
 
-        User user1 = new User();
-        user1.setName("Иван Иванов");
-        user1.setEmail("ivan@mail.ru");
-        user1.setPassword("password123");
-        userRepository.save(user1);
+            User user1 = new User();
+            user1.setName("Иван Иванов");
+            user1.setEmail("ivan@mail.ru");
+            user1.setPassword("password123");
+            userRepository.save(user1);
 
-        User user2 = new User();
-        user2.setName("Мария Петрова");
-        user2.setEmail("maria@mail.ru");
-        user2.setPassword("securepass");
-        userRepository.save(user2);
+            User user2 = new User();
+            user2.setName("Мария Петрова");
+            user2.setEmail("maria@mail.ru");
+            user2.setPassword("securepass");
+            userRepository.save(user2);
 
-        Task task1 = new Task();
-        task1.setName("Купить продукты");
-        task1.setDescription("Молоко, хлеб, яйца");
-        task1.setCompleted(false);
-        taskRepository.save(task1);
+            Task task1 = new Task();
+            task1.setName("Купить продукты");
+            task1.setDescription("Молоко, хлеб, яйца");
+            task1.setCompleted(false);
+            taskRepository.save(task1);
 
-        Task task2 = new Task();
-        task2.setName("Написать отчет");
-        task2.setDescription("Подготовить ежемесячный отчет для руководства");
-        task2.setCompleted(false);
-        taskRepository.save(task2);
+            Task task2 = new Task();
+            task2.setName("Написать отчет");
+            task2.setDescription("Подготовить ежемесячный отчет для руководства");
+            task2.setCompleted(false);
+            taskRepository.save(task2);
 
-        Task task3 = new Task();
-        task3.setName("Позвонить другу");
-        task3.setDescription("Обсудить планы на выходные");
-        task3.setCompleted(true);
-        taskRepository.save(task3);
+            Task task3 = new Task();
+            task3.setName("Позвонить другу");
+            task3.setDescription("Обсудить планы на выходные");
+            task3.setCompleted(true);
+            taskRepository.save(task3);
+        }
     }
 }
